+++ lib/utils/utils.dart
@@ -3,3 +3,4 @@ library geoff.utils;
 export './networking/networking.dart';
 export './system/system.dart';
 export './maths/maths.dart';
+export './alarm.dart';